requirements.txt
streamlit==1.32.2
joblib==1.3.2
numpy==1.26.3
opencv-python-headless==4.9.0.80
Pillow==10.2.0
scikit-learn==1.4.0